+++ src/table/index.js
@@ -31,13 +31,15 @@ const findTables = function(section, wiki) {
   let tables = [];
   list.forEach((str) => {
     if (str) {
-      wiki = wiki.replace(str, '');
+      //also reremove a newline at the end of the table (awkward)
+      wiki = wiki.replace(str + '\n', '');
       let data = parseTable(str);
       if (data && data.length > 0) {
         tables.push(new Table(data));
       }
     }
   });
+  // console.log(wiki);
   if (tables.length > 0) {
     section.tables = tables;
   }