+++ tests/table.test.js
@@ -241,3 +241,48 @@ test('embedded-table', t => {
   t.equal(tables[1].links().length, 1, 'found another link');
   t.end();
 });
+
+test('embedded-table-2', t => {
+  var str = ` {| class="oopsie"
+  | first row
+  |-
+  | Secod row
+  {|
+  |-
+  | embed 1
+  |-
+  | embed 2
+  |}
+  |-
+  | Berlin!
+  |-
+  |}
+
+  Actual first sentence is here`;
+  var doc = wtf(str);
+  t.equal(doc.tables().length, 2, 'found both tables');
+  var text = doc.sentences(0).text();
+  t.equal('Actual first sentence is here', text, 'got proper first sentence');
+  t.end();
+});
+
+
+
+test('sortable table', t => {
+  var str = `{|class="wikitable sortable"
+  !Name and Surname!!Height
+  |-
+  |data-sort-value="Smith, John"|John Smith||1.85
+  |-
+  |data-sort-value="Ray, Ian"|Ian Ray||1.89
+  |-
+  |data-sort-value="Bianchi, Zachary"|Zachary Bianchi||1.72
+  |-
+  !Average:||1.82
+  |}`;
+  let doc = wtf(str);
+  var row=doc.tables(0).data[0]
+  t.equal(row.Height.text(),'1.85','got height')
+  t.equal(row['Name and Surname'].text(),'John Smith','got name')
+  t.end();
+});